+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### Unreleased
 
+* Link `EntitySchema` instances associated with channel to `ChannelSchema` to simplify validation
+  of graph when or if needed.
+
 ### [v6.00](https://github.com/realityforge/replicant/tree/v6.00) (2018-07-02)
 [Full Changelog](https://github.com/realityforge/replicant/compare/v5.99...v6.00)